+++ reaction_diffusion/app/main.f90
@@ -1,0 +1,11 @@
+!------------------------------------------------------------------------------
+! Contributed by Vincent Magnin, 2024-02-27
+! Last modifications: vmagnin 2024-02-27
+! MIT license
+!------------------------------------------------------------------------------
+
+program main
+  use GUI_module, only: initialize_GUI
+
+  call initialize_GUI
+end program main